refactor(products): modificação do diretório padrão dos componentes de produtos
